requirements.txt
Flask==3.0.3
python-docx==1.1.2
pypdf==5.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
